+++ examples/ffmpeg-output/src/dialog.rs
@@ -33,7 +33,7 @@ impl FfmpegOutputConfigDialog {
 
 impl eframe::App for FfmpegOutputConfigDialog {
     fn ui(&mut self, ui: &mut egui::Ui, _frame: &mut eframe::Frame) {
-        egui::CentralPanel::default().show_inside(ui, |ui| {
+        egui::CentralPanel::default().show(ui, |ui| {
             egui::Grid::new("main-grid")
                 .min_col_width(ui.available_width() / 2.0)
                 .min_row_height(ui.available_height())