+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Breaking**: 依存関係を更新
 - generic: `ReadSection::get_object_section_ranges`を追加
 - generic: serdeのシリアライズにzlib-rsを使用するように変更
+- generic: `EffectHandle`にSend・Syncを実装
 
 ### デモプラグイン