+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+（なし）
+
+### デモプラグイン
+
+（なし）
+
+## [0.39.0](https://github.com/sevenc-nanashi/aviutl2-rs/releases/tag/0.39.0)
+
 - **Breaking**: 最小サポートを2.1.0に引き上げ
 - **Breaking**: 依存関係を更新
 - generic: `ReadSection::get_object_section_ranges`を追加